Adding ingredients that already exist in the fridge

+++ app/src/main/java/kitchnpal/kitchnpal/Ingredient.java
@@ -39,7 +39,7 @@ public class Ingredient {
         return this.ingredientAmount;
     }
 
-    public void setIngredientAmount(float amount) {
+    public void setIngredientAmount(double amount) {
         this.ingredientAmount = amount;
     }
